Update depth data handling and project configuration
Refactored the native depth processing pipeline to handle raw 16-bit depth data from ARCore and added missing Android XML configurations.

- **Native/C++ Changes**:
    - Updated `MobileGS::feedDepthData` and related JNI methods to use `uint16_t*` for depth pixels instead of `float*`.
    - Added conversion logic to transform raw millimeter depth values (`DEPTH16`) into meters.
    - Simplified and cleaned up internal data structures (`Splat`, `Chunk`, `ChunkKey`) by removing redundant comments.
    - Optimized depth processing loop to skip invalid data (0mm).
- **Android/Resource Changes**:
    - Added `file_paths.xml` for FileProvider support.
    - Updated `AndroidManifest.xml` to reference `@xml/file_paths`.
    - Added `backup_rules.xml`, `data_extraction_rules.xml`, and `network_security_config.xml`.
    - Cleaned up formatting in manifest.

+++ core/native/src/main/cpp/include/MobileGS.h
@@ -9,28 +9,18 @@
 #include <GLES3/gl3.h>
 #include <glm/glm.hpp>
 
-// --- SPLATAM: Augmented Data Structure ---
 struct Splat {
-    // Spatial (Geometry)
     float x, y, z;
-
-    // Orientation & Scale (Covariance approximation for Mobile)
-    float nx, ny, nz; // Surface Normal
-    float radius;     // Splat influence radius
-
-    // Visual (Appearance)
-    uint8_t r, g, b;  // Base Albedo
-
-    // Metadata
-    float confidence;       // 0.0 to 1.0
-    uint32_t lastSeenFrame; // For culling/maintenance
-    float luminance;        // Cached for lighting invariance
+    float nx, ny, nz;
+    float radius;
+    uint8_t r, g, b;
+    float confidence;
+    uint32_t lastSeenFrame;
+    float luminance;
 };
 
-// --- DISKCHUNGS: Spatial Hashing ---
 struct ChunkKey {
     int x, y, z;
-
     bool operator==(const ChunkKey& other) const {
         return x == other.x && y == other.y && z == other.z;
     }
@@ -38,7 +28,6 @@ struct ChunkKey {
 
 struct ChunkKeyHash {
     std::size_t operator()(const ChunkKey& k) const {
-        // Cantor pairing or simple XOR hash for 3D coords
         size_t h1 = std::hash<int>{}(k.x);
         size_t h2 = std::hash<int>{}(k.y);
         size_t h3 = std::hash<int>{}(k.z);
@@ -47,11 +36,9 @@ struct ChunkKeyHash {
 };
 
 struct Chunk {
-    bool isDirty;       // Needs GL buffer update
-    bool isActive;      // Is currently in memory/renderable
+    bool isDirty;
+    bool isActive;
     std::vector<Splat> splats;
-
-    // GL Buffers for this chunk
     GLuint vbo;
     int splatCount;
 
@@ -64,47 +51,39 @@ public:
     ~MobileGS();
 
     // --- Core Pipeline (SplaTAM) ---
-    void feedDepthData(const float* depthPixels, const float* colorPixels,
+    // CHANGED: depthPixels is now uint16_t* (raw millimeters)
+    void feedDepthData(const uint16_t* depthPixels, const float* colorPixels,
             int width, int height, const float* cameraPose, float fov);
 
-    void update(const float* cameraPose); // Manages chunks (load/unload)
-
-    // Renders using stored matrices or passed ones
+    void update(const float* cameraPose);
     void render(const float* viewMatrix, const float* projMatrix);
 
-    // --- JNI Adapter Methods ---
-    void initialize(); // Reset/Init
-    void updateCamera(const float* view, const float* proj); // Stores matrices for draw()
-    void draw(); // Calls render() with stored matrices
+    void initialize();
+    void updateCamera(const float* view, const float* proj);
+    void draw();
 
     void onSurfaceChanged(int width, int height);
     int getSplatCount();
     void clear();
 
-    // Serialization
     bool saveModel(std::string path);
     bool loadModel(std::string path);
     void alignMap(const float* transform);
 
-    // Settings
     void setChunkSize(float meters) { mChunkSize = meters; }
 
 private:
-    // Parameters
     float mChunkSize = 2.0f;
     float mVoxelSize = 0.05f;
     int mFrameCount = 0;
 
-    // State
     glm::mat4 mStoredView;
     glm::mat4 mStoredProj;
     int mScreenWidth, mScreenHeight;
 
-    // Storage
     std::unordered_map<ChunkKey, Chunk, ChunkKeyHash> mChunks;
     std::mutex mChunkMutex;
 
-    // Helpers
     ChunkKey getChunkKey(float x, float y, float z);
     float getLuminance(uint8_t r, uint8_t g, uint8_t b);
     void fuseSplat(Splat& target, const Splat& source);
